refactor(procedure): refactor procedure publish methods
