+++ app/controllers/administrateurs/procedures_controller.rb
@@ -307,10 +307,6 @@ module Administrateurs
 
       @procedure.publish_or_reopen!(current_administrateur)
 
-      if @procedure.draft_changed?
-        @procedure.publish_revision!
-      end
-
       if params[:old_procedure].present? && @procedure.errors.empty?
         current_administrateur
           .procedures
