+++ app/models/procedure.rb
@@ -8,6 +8,7 @@ class Procedure < ApplicationRecord
   include ProcedureGroupeInstructeurAPIHackConcern
   include ProcedureSVASVRConcern
   include ProcedureChorusConcern
+  include ProcedurePublishConcern
   include PiecesJointesListConcern
   include ColumnsConcern
 
@@ -327,39 +328,6 @@ class Procedure < ApplicationRecord
 
   def dossiers_close_to_expiration
     dossiers.close_to_expiration.count
-  end
-
-  def publish_or_reopen!(administrateur)
-    Procedure.transaction do
-      if brouillon?
-        reset!
-        cleanup_types_de_champ_options!
-      end
-
-      other_procedure = other_procedure_with_path(path)
-      if other_procedure.present? && administrateur.owns?(other_procedure)
-        other_procedure.unpublish!
-        publish!(other_procedure.canonical_procedure || other_procedure)
-      else
-        publish!
-      end
-    end
-  end
-
-  def reset!
-    if !locked? || draft_changed?
-      dossier_ids_to_destroy = draft_revision.dossiers.ids
-      if dossier_ids_to_destroy.present?
-        Rails.logger.info("Resetting #{dossier_ids_to_destroy.size} dossiers on procedure #{id}: #{dossier_ids_to_destroy}")
-        draft_revision.dossiers.destroy_all
-      end
-    end
-  end
-
-  def cleanup_types_de_champ_options!
-    draft_revision.types_de_champ.each do |type_de_champ|
-      type_de_champ.update!(options: type_de_champ.clean_options)
-    end
   end
 
   def suggested_path(administrateur)
@@ -770,23 +738,6 @@ class Procedure < ApplicationRecord
     "Procedure;#{id}"
   end
 
-  def create_new_revision(revision = nil)
-    transaction do
-      new_revision = (revision || draft_revision)
-        .deep_clone(include: [:revision_types_de_champ])
-        .tap { |revision| revision.published_at = nil }
-        .tap(&:save!)
-
-      move_new_children_to_new_parent_coordinate(new_revision)
-
-      # they are not aware of the new tdcs
-      new_revision.types_de_champ_public.reset
-      new_revision.types_de_champ_private.reset
-
-      new_revision
-    end
-  end
-
   def average_dossier_weight
     if dossiers.termine.any?
       dossiers_sample = dossiers.termine.limit(100)
@@ -798,32 +749,6 @@ class Procedure < ApplicationRecord
       MIN_WEIGHT + total_size / dossiers_sample.length
     else
       nil
-    end
-  end
-
-  def publish_revision!
-    reset!
-    cleanup_types_de_champ_options!
-    transaction do
-      self.published_revision = draft_revision
-      self.draft_revision = create_new_revision
-      save!(context: :publication)
-      published_revision.touch(:published_at)
-    end
-    dossiers
-      .state_not_termine
-      .find_each(&:rebase_later)
-  end
-
-  def reset_draft_revision!
-    if published_revision.present? && draft_changed?
-      reset!
-      transaction do
-        draft_revision.types_de_champ.filter(&:only_present_on_draft?).each(&:destroy)
-        draft_revision.update(dossier_submitted_message: nil)
-        draft_revision.destroy
-        update!(draft_revision: create_new_revision(published_revision))
-      end
     end
   end
 
@@ -863,45 +788,6 @@ class Procedure < ApplicationRecord
       self.update(estimated_dossiers_count: self.dossiers.visible_by_administration.count,
                 dossiers_count_computed_at: now)
     end
-  end
-
-  def move_new_children_to_new_parent_coordinate(new_draft)
-    children = new_draft.revision_types_de_champ
-      .includes(parent: :type_de_champ)
-      .where.not(parent_id: nil)
-    coordinates_by_stable_id = new_draft.revision_types_de_champ
-      .includes(:type_de_champ)
-      .index_by(&:stable_id)
-
-    children.each do |child|
-      child.update!(parent: coordinates_by_stable_id.fetch(child.parent.stable_id))
-    end
-    new_draft.reload
-  end
-
-  def before_publish
-    assign_attributes(closed_at: nil, unpublished_at: nil)
-  end
-
-  def after_publish(canonical_procedure = nil)
-    self.canonical_procedure = canonical_procedure
-    self.published_revision = draft_revision
-    self.draft_revision = create_new_revision
-    save!(context: :publication)
-    touch(:published_at)
-    published_revision.touch(:published_at)
-  end
-
-  def after_republish(canonical_procedure = nil)
-    touch(:published_at)
-  end
-
-  def after_close
-    touch(:closed_at)
-  end
-
-  def after_unpublish
-    touch(:unpublished_at)
   end
 
   def update_juridique_required
